Removing dtype=int64 in example documentation

+++ sklearn/svm/classes.py
@@ -1171,7 +1171,7 @@ class OneClassSVM(BaseLibSVM, OutlierMixin):
     >>> X = [[0], [0.44], [0.45], [0.46], [1]]
     >>> clf = OneClassSVM(gamma='auto').fit(X)
     >>> clf.predict(X)
-    array([-1,  1,  1,  1, -1], dtype=int64)
+    array([-1,  1,  1,  1, -1])
     >>> clf.score_samples(X)  # doctest: +ELLIPSIS
     array([1.7798..., 2.0547..., 2.0556..., 2.0561..., 1.7332...])
     """
